requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==37.0.4
Django==3.2.18
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
pycparser==2.21
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
python-dotenv==1.0.1
drf-firebase-auth==1.0.0
psycopg2-binary==2.9.3
